# Speed up search by not invoking apm