--- a/lib/install-panel.js
+++ b/lib/install-panel.js
@@ -206,12 +206,8 @@ export default class InstallPanel {
     this.refs.searchMessage.textContent = `Searching ${this.searchType} for \u201C${query}\u201D\u2026`
     this.refs.searchMessage.style.display = ''
 
-    const opts = {}
-    opts[this.searchType] = true
-    opts['sortBy'] = 'downloads'
-
     try {
-      let packages = (await this.packageManager.search(query, opts)) || []
+      const packages = (await this.client.search(query, this.searchType)) || []
       this.refs.resultsContainer.innerHTML = ''
       this.refs.searchMessage.style.display = 'none'
       if (packages.length === 0) {