--- a/lib/install-panel.js
+++ b/lib/install-panel.js
@@ -3,7 +3,6 @@
 
 import path from 'path'
 import electron from 'electron'
-import _ from 'underscore-plus'
 import {CompositeDisposable, TextEditor} from 'atom'
 import etch from 'etch'
 
@@ -233,7 +232,7 @@ export default class InstallPanel {
   }
 
   highlightExactMatch (container, query, packages) {
-    const exactMatch = _.filter(packages, pkg => pkg.name === query)[0]
+    const exactMatch = packages.filter(pkg => pkg.name === query)[0]
 
     if (exactMatch) {
       this.addPackageCardView(container, this.getPackageCardView(exactMatch))
@@ -242,7 +241,7 @@ export default class InstallPanel {
   }
 
   addCloseMatches (container, query, packages) {
-    const matches = _.filter(packages, pkg => pkg.name.indexOf(query) >= 0)
+    const matches = packages.filter(pkg => pkg.name.indexOf(query) >= 0)
 
     for (let pack of matches) {
       this.addPackageCardView(container, this.getPackageCardView(pack))