--- a/lib/install-panel.js
+++ b/lib/install-panel.js
@@ -206,8 +206,11 @@ export default class InstallPanel {
     this.refs.searchMessage.textContent = `Searching ${this.searchType} for \u201C${query}\u201D\u2026`
     this.refs.searchMessage.style.display = ''
 
+    const options = {sort: {downloads: 'desc'}}
+    options[this.searchType] = true
+
     try {
-      const packages = (await this.client.search(query, this.searchType)) || []
+      const packages = (await this.client.search(query, options)) || []
       this.refs.resultsContainer.innerHTML = ''
       this.refs.searchMessage.style.display = 'none'
       if (packages.length === 0) {