--- a/lib/install-panel.js
+++ b/lib/install-panel.js
@@ -206,7 +206,7 @@ export default class InstallPanel {
     this.refs.searchMessage.textContent = `Searching ${this.searchType} for \u201C${query}\u201D\u2026`
     this.refs.searchMessage.style.display = ''
 
-    const options = {sort: {downloads: 'desc'}}
+    const options = {}
     options[this.searchType] = true
 
     try {