--- a/lib/install-panel.js
+++ b/lib/install-panel.js
@@ -231,7 +231,7 @@ export default class InstallPanel {
   }
 
   highlightExactMatch (container, query, packages) {
-    const exactMatch = packages.filter(pkg => pkg.name === query)[0]
+    const exactMatch = packages.filter(pkg => pkg.name.toLowerCase() === query)[0]
 
     if (exactMatch) {
       this.addPackageCardView(container, this.getPackageCardView(exactMatch))
@@ -240,7 +240,7 @@ export default class InstallPanel {
   }
 
   addCloseMatches (container, query, packages) {
-    const matches = packages.filter(pkg => pkg.name.indexOf(query) >= 0)
+    const matches = packages.filter(pkg => pkg.name.toLowerCase().indexOf(query) >= 0)
 
     for (let pack of matches) {
       this.addPackageCardView(container, this.getPackageCardView(pack))