--- a/package.json
+++ b/package.json
@@ -17,7 +17,7 @@
     "glob": "4.3.1",
     "hosted-git-info": "^2.1.4",
     "marked": "^0.3.6",
-    "request": "^2.40",
+    "request": "^2.83.0",
     "roaster": "^1.1.2",
     "season": "^6.0.2",
     "semver": "^5.3.0",