--- a/spec/install-panel-spec.coffee
+++ b/spec/install-panel-spec.coffee
@@ -60,7 +60,7 @@ describe 'InstallPanel', ->
 
   describe "searching packages", ->
     it "highlights exact name matches", ->
-      spyOn(@packageManager, 'search').andCallFake ->
+      spyOn(@panel.client, 'search').andCallFake ->
         new Promise (resolve, reject) -> resolve([ {name: 'not-first'}, {name: 'first'} ])
       spyOn(@panel, 'getPackageCardView').andCallThrough()
 
@@ -72,7 +72,7 @@ describe 'InstallPanel', ->
         expect(@panel.getPackageCardView.argsForCall[1][0].name).toEqual 'not-first'
 
     it "prioritizes partial name matches", ->
-      spyOn(@packageManager, 'search').andCallFake ->
+      spyOn(@panel.client, 'search').andCallFake ->
         new Promise (resolve, reject) -> resolve([ {name: 'something else'}, {name: 'partial name match'} ])
       spyOn(@panel, 'getPackageCardView').andCallThrough()