### Description of the Change

This pull request implements the search capability equivalent to previously used `apm search --json (--packages|--themes)? term`, sorted by download count in descending order. Measuring the web request that is done inside `apm`, with ~118ms network latency it takes around 750-800ms to respond (according to Chrome devtools), whereas the total execution time for `apm search` takes around 1.5 seconds.

Searching for `github` and measuring using `performance.now()`:
Before: `packageManager.search()` 1587ms average
After: `client.search()` 853ms average

Although using just `apm` would be cleaner, this is a significant gain in terms of perceived responsiveness. Search functionality doesn't depend on `apm` existing in any way and being one of the most frequently used parts of settings-view where user has to wait, it makes sense to reduce any delays where possible. Ideally there should be a library which both `settings-view` and `apm` use as that eliminates the duplication concern.

It is possible that by enabling compression at atom.io, the request response time can be further improved, making responses a fair amount faster on slower connections as huge amounts of text typically compress well.

### Alternate Designs

None considered.

### Benefits

Search returns faster ⚡️, doesn't feel as slow to the user.

### Possible Drawbacks

A little bit more code duplication between `apm` and this package.

### Applicable Issues

None.