@50Wliu 2017-12-11
It looks like `install-panel` is the only place that calls `packageManager.search`?  Therefore it should be safe to :fire: that entire method.