@50Wliu 2018-01-02
I have absolutely no clue why apm decided to alter the returned JSON from its own API to make `description` as a top-level property but it's _really_ annoying (why not just change it in the API??).  

Anyway, the API allows for a `sort` parameter so you can pass that directly in the query string rather than sorting the data afterwards.  You can also get request to process the JSON for you.

Here's the version that I wrote while investigating a different issue.  I think the only thing that's missing from this one is the changing of `description`.
```coffee
    new Promise (resolve, reject) ->
      qs = {q: query}

      if options.themes
        qs.filter = 'theme'
      else if options.packages
        qs.filter = 'package'

      if options.sortBy
        qs.sort = options.sortBy

      options = {
        url: "https://atom.io/api/packages/search"
        headers: {'User-Agent': navigator.userAgent}
        qs: qs
        json: true
      }

      request options, (err, res, body) ->
        if err
          error = new Error("Searching for \u201C#{query}\u201D failed.")
          error.stderr = err.message
          reject(error)
        else
          resolve(body)
```