@Ingramz 2018-01-04
@50Wliu thank you for the feedback. I made some changes based on your version, should look nicer now.