@50Wliu 2018-01-08
I don't think we are sorting by downloads right now, either.  Here's what I see after I search for `random` and scroll down a bit.
![settings-view-search-random](https://user-images.githubusercontent.com/2766036/34656254-feeeb996-f3e4-11e7-9e48-b3afa487c64f.png)