@Ingramz 2018-01-08
Packages containing the search term will be moved to the top with the exact match always being the first one if there is one. This is done later in the process.

https://github.com/atom/settings-view/blob/9315802415a8d4c879bb6a2059eacd86b7d26639/lib/install-panel.js#L221-L223