@Ingramz 2018-01-09
Gzip compression was enabled 🎉 . Seeing additional ~100ms improvement, but I will try rerunning the measurements from the opening post to get more accurate numbers. Currently on a moving bus with 4G internet, I am seeing sub-second or close-to-a-second overall response times.

It occurred to me that we can probably save a portion of time by reusing the TLS session from the first connection as it is quite expensive to obtain it every time via handshake. If reuse is possible using `request` or some other method, we could also port featured packages to use the API directly from Atom. While the user is typing, the session can be fetched via the already open/completed featured packages request and the first search could be at least one RTT faster. I would not want to roll our own http client though if there is no easy way of doing this.